refactor: renamed 'BinarySearchTests.kt' to 'BinarySearchTreeTests.kt'
